Extract internal API to a submodule (#7)
